install.py changes
We've been working a very long time on trying to fix install.py.  There has been constant misunderstanding by Claude on what needs to be done and just did a massive refactoring of setup. instead of install.py.  Nonetheless some of the fixes were good The major issue was that when the installer finished it would open a browser as a choice which does not help remote installations so we remove that and also it redirected the user on localhost to /login when it should go to /setup. After many thousands of wasted tokens and time Claude was finally able to figure this out but it took over 5 hours.